Disable Go modules and cgo
When the smoke tests release was updated to Go v1.11.2, the BOSH release
pipelines all passed, but pipelines that consumed the final release
failed because `gcc` was not installed.  This is a very similar issue to
https://github.com/golang/go/issues/26988.  Experiments showed that
adding `CGO_ENABLED=0 GO111MODULE=off` to disable `cgo` and Go moudles
allowed the smoke tests to pass.  These variables should be removed when
we recieved an updated version of Go (after 1.11.2).

[#159658309]