Make smoke tests log to a file
From Bosh.io docs: https://bosh.io/docs/errands/

> Errand output is limited to one megabyte of data.
> If the response is larger an error will be returned instead.
> Release authors should ensure the full error is properly logged
> to disk and a summary is returned as output.

This commit makes smoke tests log to a file, in addition to the standard
output. Now, the standard output is a small summary, and the log file
has every single line captured by Ginkgo writer.

This commit also changes the step where we get a service key, so that it
does not print the service key. The service key has credentials that are
temporary for the smoke tests, however, we should not log credentials to
any output.

Signed-off-by: Aitor Perez Cedres <acedres@vmware.com>